+++ src/image-ops/index.js
@@ -1,5 +1,6 @@
 const size = require('./size')
 const stylize = require('./stylize')
+const settings = require('../helpers/settings')
 
 const operationsByCategory = {
   size: size.apply,
@@ -30,12 +31,12 @@ exports.apply = async (image, edits) => {
 }
 
 exports.restrictSize = async (image, metadata) => {
-  const maxImgWidth = process.env.MAX_IMAGE_WIDTH !== null ? parseInt(process.env.MAX_IMAGE_WIDTH) : null;
-  const maxImgHeight = process.env.MAX_IMAGE_HEIGHT !== null ? parseInt(process.env.MAX_IMAGE_HEIGHT) : null;
+  const maxImgWidth = settings.getSetting('MAX_IMAGE_WIDTH')
+  const maxImgHeight = settings.getSetting('MAX_IMAGE_HEIGHT')
   if ((maxImgWidth && metadata.width > maxImgWidth) || (maxImgHeight && metadata.height > maxImgHeight)) {
     const aspectRatio = parseFloat(metadata.width) / metadata.height
-    const width = aspectRatio >= 1 ? maxImgWidth : null,
-        height = width === null ? maxImgHeight : null
+    const width = aspectRatio >= 1 ? maxImgWidth : null
+    const height = width === null ? maxImgHeight : null
     await size.scaleMax(image, width, height)
   }
 }