+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `SLS_VALID_PATH_REGEX` setting to explicitly control what paths are allowed (Thanks @bs-thomas) (#52)
 - Add support for deploying to custom region (#61)
 - Added support for SSL & custom domain (Thanks @bs-thomas) (#47)
+- Added `DEFAULT_CACHE_CONTROL` to allow settings a cache control value when one isn't set on the source object
+- Added `settings.js` helper
 
 ### Fixed
 - Error that could occur when cropping via focalpoint and a non-integer is encountered